server/requirements.txt
certifi==2020.6.20
chardet==4.0.0
click==8.1.3
colorzero==1.1
distro==1.5.0
Flask==2.2.2
gpiozero==1.6.2
idna==2.10
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.19.5
picamera2==0.3.7
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
Pillow==8.1.2
python-apt==2.2.1
python-prctl==1.7
requests==2.25.1
RPi.GPIO==0.7.0
simplejpeg==1.6.4
six==1.16.0
spidev==3.5
ssh-import-id==5.10
toml==0.10.1
urllib3==1.26.5
v4l2-python3==0.3.1
Werkzeug==2.2.2
zipp==3.11.0
